TVGUI-17 Move BackButton component into directory
