+++ src/Pages/RecordedShow.tsx
@@ -1,10 +1,10 @@
-import React, { useState, useEffect } from 'react';
+import { useState, useEffect } from 'react';
 import { useParams } from "react-router";
 import { Alert, Button, Table, TableColumnsType, Tag } from 'antd';
 
 import { getRecordedShow } from '../requests/requests';
 import { Episode, RecordedShowModel } from '../utils';
-import BackButton from '../BackButton';
+import BackButton from '../components/BackButton';
 import '../RecordedShowData.css';
 import '../EpisodeData.css';
